mohsen_backend_consumer/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
python-dotenv
Werkzeug
kafka-python
requests==2.26.0
msgpack
